[refactor] #25 - 세션 시작하기 모달 화면 뷰 파일 분리
